changing docs to cpcats, fixing typos
